+++ lib/src/nodes/container_node.dart
@@ -1,10 +1,8 @@
 import 'dart:html' as html;
 
 import 'package:dawn/src/node_creator.dart';
+import 'package:dawn/src/nodes.dart';
 import 'package:dawn/src/widgets.dart';
-
-import 'framework_node.dart';
-import 'node.dart';
 
 class ContainerNode extends FrameworkNode<Container, html.DivElement> {
   late List<Node> childNodes;