+++ lib/src/nodes/framework_node.dart
@@ -1,9 +1,7 @@
 import 'dart:html' as html;
 
+import 'package:dawn/src/nodes.dart';
 import 'package:dawn/src/widgets.dart';
-
-import 'container_node.dart';
-import 'node.dart';
 
 abstract class FrameworkNode<T extends FrameworkWidget, U extends html.Element>
     extends Node<T> {