+++ lib/src/nodes/textbox_node.dart
@@ -1,8 +1,7 @@
 import 'dart:html' as html;
 
+import 'package:dawn/src/nodes.dart';
 import 'package:dawn/src/widgets.dart';
-
-import 'user_input_node.dart';
 
 class TextBoxNode extends UserInputNode<Textbox, html.TextAreaElement> {
   TextBoxNode(super.widget, {super.parentNode})