+++ lib/src/nodes/user_input_node.dart
@@ -1,8 +1,7 @@
 import 'dart:html' as html;
 
+import 'package:dawn/src/nodes.dart';
 import 'package:dawn/src/widgets.dart';
-
-import 'framework_node.dart';
 
 abstract class UserInputNode<T extends UserInputWidget, U extends html.Element>
     extends FrameworkNode<T, U> {