+++ lib/src/widgets/image.dart
@@ -1,4 +1,4 @@
-import 'framework_widget.dart';
+import 'package:dawn/src/widgets.dart';
 
 class Image extends FrameworkWidget {
   final String source;