+++ lib/src/widgets/stateful_widget.dart
@@ -1,6 +1,5 @@
 import 'package:dawn/src/nodes.dart';
-
-import 'widget.dart';
+import 'package:dawn/src/widgets.dart';
 
 abstract class StatefulWidget extends Widget {
   const StatefulWidget({super.key});