+++ lib/src/widgets/stateless_widget.dart
@@ -1,5 +1,4 @@
-import 'buildable.dart';
-import 'widget.dart';
+import 'package:dawn/src/widgets.dart';
 
 abstract class StatelessWidget extends Widget with Buildable {
   const StatelessWidget({super.key});